authorization part (user)

+++ backend/routes/post.js
@@ -33,7 +33,8 @@ router.post("", checkAuth, upload.single("image"), function(req, res, next) {
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    author: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -56,11 +57,18 @@ router.put("/:id", checkAuth, upload.single("image"), function(req, res, next) {
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    author: req.userData.userId
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "updated" });
+  Post.updateOne({ _id: req.params.id, author: req.userData.userId }, post).then(result => {
+    if(result.nModified > 0){
+      res.status(200).json({ message: "updated" });
+    }
+    else{
+      res.status(401).json({ message: "Not Authorized" });
+    }
+    //console.log(result);
+
   });
 });
 // rwdavid // ycf9jAk57srl6cm0
@@ -98,11 +106,14 @@ router.get("/:id", function(req, res, next) {
 });
 
 router.delete("/:id", checkAuth, function(req, res, next) {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(req.params.id);
-    res.status(200).json({
-      message: "Post deleted succesfully"
-    });
+  Post.deleteOne({ _id: req.params.id, author: req.userData.userId }).then(result => {
+    //console.log(req.params.id);
+    if(result.n > 0){
+      res.status(200).json({ message: "Deleted" });
+    }
+    else{
+      res.status(401).json({ message: "Not Authorized" });
+    }
   });
 });
